requirements.txt
discord.py
twitch-python
pickledb